Abstracted authenticate for routes and file upload function
